added long long to previous_timestamp

+++ src/FusionEKF.h
@@ -36,7 +36,7 @@ private:
   bool is_initialized_;
 
   // previous timestamp
-  long previous_timestamp_;
+  long long previous_timestamp_;
 
   float noise_ax;
   float noise_ay;
